+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - simplify the calculation of cluster stats
 - class/cluster numbering starts at 1 (0-based in autoclass output)
 - add dendrogram of classes
+- add reproducible run option (test only)
 
 **0.1.12**
 - update write_cluster_stats() methods with mean/std column